% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkWidget.R
\name{vtkImageData}
\alias{vtkImageData}
\title{vtk Image Data}
\usage{
vtkImageData(
  a,
  origin = rep(0, length(dim(a))),
  spacing = rep(1, length(dim(a)))
)
}
\arguments{
\item{a}{3d array}

\item{origin}{image origin}

\item{spacing}{image spacing}
}
\description{
Convenience function to create a list
object for an image that can
be passed to the vtkWidget. The vtkWidget
will pass it to the unerying
java script as a javascript object
}
