% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkVolumeRender.R
\name{vtkVolumeRender-shiny}
\alias{vtkVolumeRender-shiny}
\alias{vtkVolumeRenderOutput}
\alias{renderVtkVolumeRender}
\title{Shiny bindings for vtkVolumeRender}
\usage{
vtkVolumeRenderOutput(outputId, width = "100\%", height = "400px")

renderVtkVolumeRender(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a vtkVolumeRender}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using vtkVolumeRender within Shiny
applications and interactive Rmd documents.
}
\examples{

library(vtkwidgets)
library(shiny)

runApp(list(
 ui = bootstrapPage(
   actionButton("update","update gauge"),
   # example use of the automatically generated output function
   vtkVolumeRenderOutput("volume1")
 ),
 server = function(input, output) {

   # reactive that generates a random value for the gauge
   value = reactive({
     input$update
     array(runif(1000), dim=rep(10, 3) )
   })

   # example use of the automatically generated render function
   output$volume1 <- renderVtkVolumeRender({
     vtkVolumeRender(value())
   })
 }
))

}
