% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkVolumeRender.R
\name{vtkVolumeRender}
\alias{vtkVolumeRender}
\title{<Add Title>}
\usage{
vtkVolumeRender(image, width = NULL, height = NULL, elementId = NULL)
}
\description{
<Add Description>
}
\examples{

library(vtkwidgets)
vtkVolumeRender( array(runif(1000), dim=rep(10, 3) ) )

}
