% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkWidget.R
\name{vtkWidget-shiny}
\alias{vtkWidget-shiny}
\alias{vtkWidgetOutput}
\alias{renderVtkWidget}
\title{Shiny bindings for vtkWidget}
\usage{
vtkWidgetOutput(outputId, width = "100\%", height = "400px")

renderVtkWidget(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a vtkWidget}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using vtkWidget within Shiny
applications and interactive Rmd documents.
}
\examples{

library(vtkwidgets)
library(shiny)

runApp(list(
 ui = bootstrapPage(
   actionButton("update","update gauge"),
   # example use of the automatically generated output function
   vtkWidgetOutput("volume1")
 ),
 server = function(input, output) {

   # reactive that generates a random value for the gauge
   value = reactive({
     input$update
     array(runif(1000), dim=rep(10, 3) )
   })

   # example use of the automatically generated render function
   output$volume1 <- renderVtkWidget({
     vtkWidget(value())
   })
 }
))

}
