% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkWidget.R
\name{vtkWidget}
\alias{vtkWidget}
\title{<Add Title>}
\usage{
vtkWidget(
  data,
  renderFunction = "function(x, global){}",
  resizeFunction = "function(width, height, global){}",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
\examples{

library(vtkwidgets)

vtkWidget()

}
