% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkWidget.R
\name{vtkVolumeRender}
\alias{vtkVolumeRender}
\title{vtk Volume Rendering}
\usage{
vtkVolumeRender(image, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{image}{3d array to volume render}
}
\description{
Convenience wrapper to create vtkWidget for
volume rendering
}
\examples{

library(vtkwidgets)
vtkVolumeRender( array(runif(1000), dim=rep(10, 3) ) )

}
